Refactoring, fix build issues
